requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Deep Learning (Optional - for LSTM)
tensorflow>=2.13.0
keras>=2.13.0

# Time Series
statsmodels>=0.14.0
prophet>=1.1.4

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Explainability
shap>=0.42.0

# Signal Processing
pywavelets>=1.4.1

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0